backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
openai
jinja2
google-generativeai
pymongo
langchain-huggingface
langchain-text-splitters
langchain-mongodb
langchain-core
sentence-transformers
pypdf

json_repair
python-multipart
playwright